# resources/Construcao-Software-Swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Construcao-Software",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://192.168.0.8:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/students/{studentid}/enrolls": {
      "get": {
        "tags": [
          "enrolls"
        ],
        "summary": "Get All",
        "operationId": "GetAll1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAll1"
                  },
                  "description": "",
                  "example": [
                    {
                      "_id": "614b8a902b8404587132d054",
                      "semester": "5",
                      "classId": "614b8a902b8404587132d053",
                      "studentId": "614b8a882b8404587132d04f"
                    }
                  ]
                },
                "example": [
                  {
                    "_id": "614b8a902b8404587132d054",
                    "semester": "5",
                    "classId": "614b8a902b8404587132d053",
                    "studentId": "614b8a882b8404587132d04f"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "post": {
        "tags": [
          "enrolls"
        ],
        "summary": "Register",
        "operationId": "PostRegister",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest1"
              },
              "example": {
                "semester": "5",
                "classId": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegisterMissingValue1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Invalid value",
                      "param": "classId",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Invalid value",
                    "param": "classId",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
    "/api/students/{studentid}/enrolls/{enrollid}": {
      "get": {
        "tags": [
          "enrolls"
        ],
        "summary": "Get One",
        "operationId": "GetOne1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enrollid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOne1"
                },
                "example": {
                  "_id": "614b8a902b8404587132d054",
                  "semester": "5",
                  "classId": "614b8a902b8404587132d053",
                  "studentId": "614b8a882b8404587132d04f"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetOneWrongFormat1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Cast to ObjectId failed for value \"fhcgvjkhbjnkm\" (type string) at path \"_id\" for model \"Enrolls\""
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Cast to ObjectId failed for value \"fhcgvjkhbjnkm\" (type string) at path \"_id\" for model \"Enrolls\""
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetOneWrongID1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "enroll not found with id=614b7defed98c3e7891ec4de and studentId=614b8a882b8404587132d04f"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "enroll not found with id=614b7defed98c3e7891ec4de and studentId=614b8a882b8404587132d04f"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "put": {
        "tags": [
          "enrolls"
        ],
        "summary": "Total Update",
        "operationId": "PutTotalUpdate",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enrollid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotalUpdateRequest1"
              },
              "example": {
                "semester": "6",
                "classId": "",
                "studentId": "{{studentid}}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TotalUpdateMissingValue1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Invalid value",
                      "param": "classId",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Invalid value",
                    "param": "classId",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "patch": {
        "tags": [
          "enrolls"
        ],
        "summary": "Partial Update",
        "operationId": "PatchPartialUpdate",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enrollid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialUpdateRequest1"
              },
              "example": {
                "semester": "7"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartialUpdateWrongType3"
                  },
                  "description": "",
                  "example": [
                    {
                      "value": 5,
                      "msg": "Invalid value",
                      "param": "semester",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "value": 5,
                    "msg": "Invalid value",
                    "param": "semester",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "delete": {
        "tags": [
          "enrolls"
        ],
        "summary": "Remove",
        "operationId": "DeleteRemove",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enrollid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoveWrongFormat1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Cast to ObjectId failed for value \"fhcgvjkhbjnkm\" (type string) at path \"_id\" for model \"Enrolls\""
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Cast to ObjectId failed for value \"fhcgvjkhbjnkm\" (type string) at path \"_id\" for model \"Enrolls\""
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoveWrongID1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "enroll not found with id=614b7defed98c3e7891ec4de and studentId=614b8a882b8404587132d04f"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "enroll not found with id=614b7defed98c3e7891ec4de and studentId=614b8a882b8404587132d04f"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetOneWrongFormat1": {
        "title": "GetOneWrongFormat1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Cast to ObjectId failed for value \"drftgyvbhjn\" (type string) at path \"_id\" for model \"Students\""
        }
      },
      "GetOneWrongID1": {
        "title": "GetOneWrongID1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "student not found with id=614b7defed98c3e7891ec4de"
        }
      },
...
      "RegisterMissingValue1": {
        "title": "RegisterMissingValue1",
        "required": [
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Invalid value",
          "param": "birthday",
          "location": "body"
        }
      },
...
      "TotalUpdateMissingValue1": {
        "title": "TotalUpdateMissingValue1",
        "required": [
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Invalid value",
          "param": "birthday",
          "location": "body"
        }
      },
...
      "RemoveWrongFormat1": {
        "title": "RemoveWrongFormat1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Cast to ObjectId failed for value \"kll\" (type string) at path \"_id\" for model \"Students\""
        }
      },
      "RemoveWrongID1": {
        "title": "RemoveWrongID1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "student not found with id=614b7deced98c3e7891ec4de"
        }
      },
      "GetAll1": {
        "title": "GetAll1",
        "required": [
          "_id",
          "semester",
          "classId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "semester": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "example": {
          "_id": "614b8a902b8404587132d054",
          "semester": "5",
          "classId": "614b8a902b8404587132d053",
          "studentId": "614b8a882b8404587132d04f"
        }
      },
      "GetOne1": {
        "title": "GetOne1",
        "required": [
          "_id",
          "semester",
          "classId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "semester": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "example": {
          "_id": "614b8a902b8404587132d054",
          "semester": "5",
          "classId": "614b8a902b8404587132d053",
          "studentId": "614b8a882b8404587132d04f"
        }
      },
      "RegisterRequest1": {
        "title": "RegisterRequest1",
        "required": [
          "semester",
          "classId"
        ],
        "type": "object",
        "properties": {
          "semester": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          }
        },
        "example": {
          "semester": "5",
          "classId": ""
        }
      },
...
      "TotalUpdateRequest1": {
        "title": "TotalUpdateRequest1",
        "required": [
          "semester",
          "classId",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "semester": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "example": {
          "semester": "6",
          "classId": "",
          "studentId": "{{studentid}}"
        }
      },
...
      "PartialUpdateRequest1": {
        "title": "PartialUpdateRequest1",
        "required": [
          "semester"
        ],
        "type": "object",
        "properties": {
          "semester": {
            "type": "string"
          }
        },
        "example": {
          "semester": "7"
        }
      },
      "PartialUpdateWrongType3": {
        "title": "PartialUpdateWrongType3",
        "required": [
          "value",
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "value": 5,
          "msg": "Invalid value",
          "param": "semester",
          "location": "body"
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}